Update pin initialization and usage Add passive buzzer

+++ digital_input_pin.hpp
@@ -8,7 +8,7 @@ namespace ardent
 class digital_input_pin : public pin
 {
 public:
-    digital_input_pin(int pin_number);
+    digital_input_pin(uint8_t pin_number);
 
     virtual bool is_high() const;
     virtual bool is_low() const;
